Add RecognizerId.

+++ grade/internal/domain/endorsed/endorsed_id.go
@@ -4,18 +4,18 @@ import (
 	"github.com/emacsway/qualifying-grade/grade/internal/domain/seedwork/interfaces"
 )
 
-func NewEndorsedId(value int) EndorsedId {
+func NewEndorsedId(value uint64) EndorsedId {
 	return EndorsedId{value: value}
 }
 
 type EndorsedId struct {
-	value int
+	value uint64
 }
 
-func (id EndorsedId) Equals(other interfaces.Identity[int]) bool {
+func (id EndorsedId) Equals(other interfaces.Identity[uint64]) bool {
 	return id.value == other.GetValue()
 }
 
-func (id EndorsedId) GetValue() int {
+func (id EndorsedId) GetValue() uint64 {
 	return id.value
 }
